project_status/requirements.txt
fastapi
uvicorn
structlog
google-cloud-firestore
pydantic-settings
python-dotenv
